Issue #14937: Migrated section 4.1.2 Nonempty blocks K and R style to chapter wise testing
