+++ src/it/resources/com/google/checkstyle/test/chapter4formatting/rule412nonemptyblocks/InputLeftCurlyMethod.java
@@ -1,26 +1,26 @@
 package com.google.checkstyle.test.chapter4formatting.rule412nonemptyblocks;
 
 class InputLeftCurlyMethod
-{ //warn
+{ // violation ''{' at column 1 should be on the previous line.'
     InputLeftCurlyMethod() {} //ok
     InputLeftCurlyMethod(String aOne) {//ok
     }
     InputLeftCurlyMethod(int aOne)
-    { //warn
+    { // violation ''{' at column 5 should be on the previous line.'
     }
 
     void method1() {}//ok
     void method2() {//ok
     }
     void method3()
-    { //warn
+    { // violation ''{' at column 5 should be on the previous line.'
     }
     void                                                               method4()
-    { //warn
+    { // violation ''{' at column 5 should be on the previous line.'
     }
     void method5(String aOne,
                  String aTwo)
-    {//warn
+    {// violation ''{' at column 5 should be on the previous line.'
     }
     void method6(String aOne,
                  String aTwo) {// ok
@@ -28,21 +28,21 @@ class InputLeftCurlyMethod
 }
 
 enum InputLeftCurlyMethodEnum
-{//warn
+{// violation ''{' at column 1 should be on the previous line.'
     CONSTANT1("hello")
-    {//warn
+    {// violation ''{' at column 5 should be on the previous line.'
         void method1() {}//ok
         void method2() {
         }//ok
         void method3()
-        {//warn
+        {// violation ''{' at column 9 should be on the previous line.'
         }
         void                                                               method4()
-        { //warn
+        { // violation ''{' at column 9 should be on the previous line.'
         }
         void method5(String aOne,
                      String aTwo)
-        {//warn
+        {// violation ''{' at column 9 should be on the previous line.'
         }
         void method6(String aOne,
                      String aTwo) {//ok
@@ -54,11 +54,11 @@ enum InputLeftCurlyMethodEnum
     },
 
     CONSTANT3("hellohellohellohellohellohellohellohellohellohellohellohellohellohello")
-    {//warn
+    {// violation ''{' at column 5 should be on the previous line.'
     };
 
     private InputLeftCurlyMethodEnum(String value)
-    {//warn
+    {// violation ''{' at column 5 should be on the previous line.'
 
     }
 
@@ -66,14 +66,14 @@ enum InputLeftCurlyMethodEnum
     void method2() {//ok
     }
     void method3()
-    {//warn
+    {// violation ''{' at column 5 should be on the previous line.'
     }
     void                            method4()
-    {//warn
+    {// violation ''{' at column 5 should be on the previous line.'
     }
     void method5(String aOne,
                  String aTwo)
-    {//warn
+    {// violation ''{' at column 5 should be on the previous line.'
     }
     void method6(String aOne,
                  String aTwo) {// ok
