+++ src/it/resources/com/google/checkstyle/test/chapter4formatting/rule412nonemptyblocks/InputRightCurly.java
@@ -1,9 +1,11 @@
 package com.google.checkstyle.test.chapter4formatting.rule412nonemptyblocks;
 
-public class InputRightCurly2 {
+public class InputRightCurly {
     public static void main(String[] args) {
         boolean after = false;
         try {
-        } finally { after = true; }
+        } finally {
+            after = true;
+        }
     }
 }
