+++ src/it/resources/com/google/checkstyle/test/chapter4formatting/rule412nonemptyblocks/InputRightCurlySwitchCase.java
@@ -9,14 +9,14 @@ public class InputRightCurlySwitchCase {
                 int x = 1;
                 break;
             default :
-                x = 0; } // warn
+                x = 0; } // violation ''}' at column 24 should be alone on a line.'
     }
 
     public static void method1() {
         int mode = 0;
         switch (mode) {
         default :
-               int x = 0; } // warn
+               int x = 0; } // violation ''}' at column 27 should be alone on a line.'
     }
 
     public static void method2() {
